% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignLabels.R
\name{assignLabels}
\alias{assignLabels}
\title{assignLabels}
\usage{
assignLabels(sce, reference, refColLab = "SingleR", ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{reference}{a SingleCellExperiment object for the cell types
reference}

\item{refColLab}{the column in the reference colData to use for the
cell types assignment}

\item{...}{additional arguments to pass to the method}
}
\value{
a SingleCellExperiment object with a new column in the colData
by the name of the method applied (default is SingleR)
}
\description{
assigns cell type labels by using a label assignment
approach (default is SingleR)
}
\examples{
# TBD

}
