% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildATACScores.R
\name{buildATACScores}
\alias{buildATACScores}
\title{buildATACScores}
\usage{
buildATACScores(
  sce,
  cellTypesCol = "SingleR",
  cellType,
  assayName = "ATAC",
  lncounts = FALSE
)
}
\arguments{
\item{sce}{object with cellTypesCol column in the
colData}

\item{cellTypesCol}{character indicating the colData column name to use for
selecting the barcodes}

\item{cellType}{character indicating the cell to retrieve from the
cellTypesCol}

\item{assayName}{name of the ATAC assay (default is ATAC)}

\item{lncounts}{logical for computing the score on the counts or the
lognormalized counts (logNormCouts is used), default is FALSE}
}
\value{
a SCE with a `score` column in the rowRanges for the atac exp
}
\description{
Computes a `score` for the peaks stored in the ATAC exp and
for the cell type indicated as input.
In case of multiple columns it simply computes a sum of the values for the
counts/logNormalizesCounts and stores the results in a new score column in
the rowRanges of the ATAC experiment.
}
\examples{
TBD
}
