% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{computeDoublets}
\alias{computeDoublets}
\title{computeDoublets}
\usage{
computeDoublets(sce, method = c("auto", "optim", "dbr", "griffiths"))
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{method}{one of "auto", "optim", "dbr", "griffiths", see
\link[scDblFinder]{doubletThresholding}}
}
\value{
a SingleCellExperiment object with two additional columns dbl.calls
and dbl.scores
}
\description{
computes doublets on a SingleCellExperiment object by aid of
scDblFinder package
}
\examples{
# TBD

}
