% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{computeFilterCellsMetrics}
\alias{computeFilterCellsMetrics}
\title{computeFilterCellsMetrics}
\usage{
computeFilterCellsMetrics(
  sce,
  metric = c("quantile", "isOutlier", "both"),
  lowQuantThr = "5\%",
  highQuantThr = "90\%",
  nucleosomeThr = 2,
  TSSThr = 1
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{metric}{one of quantile/isOutlier/both}

\item{lowQuantThr}{lower threshold for quantiles to label the cells to
filter out (default is 5%)}

\item{highQuantThr}{higher threshold for quantiles to label the cells to
filter out (default is 90%)}

\item{nucleosomeThr}{threshold for the nucleosome signal (default is 2)}

\item{TSSThr}{threshold for TSS enrichment (default is 1)}
}
\value{
a SingleCellExperiment with one ligical column in colData per each
metric
}
\description{
computes metrics on the measure previously computed with
computeSignacMetrics creating TRUE/FALSE columns in the sce colData
indicating if the cell passes or not the checks
}
\examples{
#TBD
}
