% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulk.R
\name{computePseudoBulk}
\alias{computePseudoBulk}
\title{computePseudoBulk}
\usage{
computePseudoBulk(sce, assayName = "counts", ctCol, ...)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{assayName}{default is counts}

\item{ctCol}{a string indicating the name of the column with the cell types
to use for the bulk transformation}

\item{...}{additional arguments to pass to \link[scuttle]{aggregateAcrossCells}}
}
\value{
a SingleCellExperiment object with pseudobulk counts
}
\description{
this function computes pseudobulk counts based on the cell types
indicated in the colData ctCol and using the assayName assay
}
\examples{
# TBD
}
