% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSignacMetrics.R
\name{computeSignacMetrics}
\alias{computeSignacMetrics}
\title{computeSignacMetrics}
\usage{
computeSignacMetrics(sce, annotation)
}
\arguments{
\item{sce}{a SingleCellExperiment object created with read10xMultiome}

\item{annotation}{a GRange object with gene annotations}
}
\value{
a SingleCellExperiment object
}
\description{
computeSignacMetrics
}
\examples{
#TBD
}
