% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBamCt.R
\name{createBamCt}
\alias{createBamCt}
\title{createBamCt}
\usage{
createBamCt(
  sce,
  cellTypesCol = "SingleR",
  cellType,
  bamdir,
  bamType = c("both", "GEX", "ATAC"),
  outdir,
  ncores = 1
)
}
\arguments{
\item{sce}{a SingleCellExperiment object with cellTypesCol column in the
colData}

\item{cellTypesCol}{character indicating the colData column name to use for
selecting the barcodes}

\item{cellType}{character indicating the cell to retrieve from the
cellTypesCol}

\item{bamdir}{path to the directory where to find the original bam files}

\item{bamType}{one of "GEX", "ATAC", "both" indicating if the user wants to
produce bams for both omics or just GEX or ATAC (default is both)}

\item{outdir}{the path to output directory}

\item{ncores}{number of cores to use}
}
\value{
none
}
\description{
starting from a cell type annotated on the SingleCellExperiment
it constructs bam files with only the barcodes of the cells annotated
for the specified cell type.
Names of the original bam files are: "gex_possorted_bam.bam$",
"atac_possorted_bam.bam$". At the moment hard-coded.
}
\examples{
TBD
}
