% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{.plotDoublets}
\alias{.plotDoublets}
\title{plotDoublets}
\usage{
.plotDoublets(sce, doubl_by, filename)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{doubl_by}{the colData column for the doublets to use for colouring the
t-SNE}

\item{filename}{the plot filename with extension}
}
\value{
a ggplot object
}
\description{
plot a tsne and saves a file with the specified filename
}
\examples{
#TBD

}
\keyword{internal}
