% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{plotDoubletsTSNEbyCT}
\alias{plotDoubletsTSNEbyCT}
\title{plotDoublets}
\usage{
plotDoubletsTSNEbyCT(
  sce,
  ctCol = "subcluster",
  doubl_by,
  cellTypes,
  path = ".",
  postfix = NULL
)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{ctCol}{the colData column with cell types}

\item{doubl_by}{the colData column for the doublets to use for colouring the
t-SNE}

\item{cellTypes}{a vector of cell types}

\item{path}{the path to save the plot}

\item{postfix}{an optional postfix for the filename}
}
\value{
none
}
\description{
plotDoublets
}
\examples{
#TBD

}
