% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCells.R
\name{plotFilteredCells}
\alias{plotFilteredCells}
\title{plotFilteredCells}
\usage{
plotFilteredCells(
  sce,
  filterCellsBy = c("all", "custom"),
  inout = c("out", "in"),
  customColName = c("in_signac", "passed_both"),
  name
)
}
\arguments{
\item{sce}{A SingleCellExperiment Object}

\item{filterCellsBy}{one of all/custom, in case of custom specify a custom
column name with customColName}

\item{inout}{one of in/out to consider the kept/removed cells}

\item{customColName}{the colData column to use to subset the SCE object to
consider}

\item{name}{the name to print on top of the plot}
}
\value{
a ggplot2 object
}
\description{
creates violin plots per each metric computed with
computeFilterCellsMetrics
}
\examples{
# TBD

}
