% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpeaks.R
\name{plotPeaksSamplesDEScan2}
\alias{plotPeaksSamplesDEScan2}
\title{plotPeaksSamplesDEScan2}
\usage{
plotPeaksSamplesDEScan2(grl, grcol = "k-carriers", title = "Peaks-samples")
}
\arguments{
\item{grl}{a GenomicRangesList with `grcol` stored in each GenomicRanges metadata
(default is `k-carriers`, as from `DEScan2::finaRegions` function)}

\item{grcol}{the column where to take the number of peaks in each GenomicRanges}

\item{title}{the title of the final plot}
}
\value{
a ggplot object
}
\description{
plotPeaksSamplesDEScan2
}
\examples{
TBD
}
