% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulk.R
\name{pseudoBulkRNA}
\alias{pseudoBulkRNA}
\title{pseudoBulkRNA}
\usage{
pseudoBulkRNA(sce, ctCol)
}
\arguments{
\item{sce}{a SingleCellExperiment object}

\item{ctCol}{a string indicating the name of the column with the cell types
to use for the bulk transformation}
}
\value{
a SingleCellExperiment object with pseudobulk counts
}
\description{
this function computes pseudobulk counts based on the cell types
indicated in the colData ctCol
}
\examples{
# TBD
}
