% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read10xMultiome.R
\name{read10xMultiome}
\alias{read10xMultiome}
\title{Load data from a 10X Genomics Multiome experiment}
\usage{
read10xMultiome(
  sample.path,
  sample.name = NULL,
  type = c("auto", "sparse", "HDF5", "prefix"),
  data = c("filtered", "raw"),
  compressed = NULL,
  col.names = FALSE,
  annotation = FALSE,
  addfrags = FALSE,
  reference = c("RNA", "ATAC")
)
}
\arguments{
\item{sample.path}{string indicating the path to the experiment outs data.
If type is sparse, it must indicate the path where the sparse data are stored.
If type is HDF5, it must indicate the name of the HDF5 file.
Alternatively, the string may contain a prefix of the name for the three-file
system described above, where the rest of the name of each file follows
the standard 10X output.
(see \link[DropletUtils]{read10xCounts} for additional information)}

\item{sample.name}{the name of the sample, when NULL is the root of the
sample.path (default is NULL)}

\item{type}{String specifying the type of 10X format to read data from.}

\item{data}{string specifying if to load filtered or raw data}

\item{compressed}{Logical scalar indicating whether the text files are
compressed for type="sparse" or "prefix".}

\item{col.names}{logical indicating if to add ID to the cols of the assays.}

\item{annotation}{boolean specifying if to load the 10x peak annotations
NB this process is very slow (default FALSE)}

\item{addfrags}{boolean indicating to store the fragments file path
the path will be stored as metadata of the ATAC assay (default FALSE)}

\item{reference}{string indicating the main assay to use for the
SingleCellExperiment object}
}
\value{
a SingleCellExperiment object with the RNA (or ATAC see
\code{reference} parameter) as main assay and ATAC (or RNA see reference
parameter) as \code{altExp} assay.
}
\description{
Creates a \linkS4class{SingleCellExperiment} from the CellRanger
ARC output directories for 10X Genomics data.
}
\examples{
# TBD see data dir

}
