% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitSCE}
\alias{splitSCE}
\title{splitSCE}
\usage{
splitSCE(sce, by, col = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{by}{the colname of the rowData or colData to refer to when splitting
the `sce`}

\item{col}{boolean (default `TRUE`) indicating if the `by` column is on the
`colData` or `rowData`}
}
\value{
a named list of SingleCellExpreriment objects
}
\description{
splitSCE
}
\examples{
TBD
}
