% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{createTracks}
\alias{createTracks}
\title{createTracks}
\usage{
createTracks(
  sce,
  cellTypesCol = "singleR",
  cellType,
  bamdir,
  outdir,
  ncores = 1
)
}
\arguments{
\item{ncores}{}
}
\description{
createTracks
}
